tool1/requirements.txt
requests==2.12.3
robotfw_resources==2.6.4
pytest==3.1.1
pytest-mockito==0.0.3
mockito==1.0.11
responses==0.8.1
